Add bootstrap script.
This is a basic bootstrap script that can be run to setup a local dev
environment. You can read through bootstrap scripts in other repos for
examples of how to make this more robust. The next thing I would add is
for it to install npm instead of just failing.